+++ classes/output/my/renderer.php
@@ -21,6 +21,7 @@ namespace tool_mucertify\output\my;
 
 use tool_mucertify\local\assignment;
 use stdClass;
+use tool_mucertify\local\certification;
 
 /**
  * My certification renderer.
@@ -56,13 +57,8 @@ class renderer extends \plugin_renderer_base {
         }
 
         $certificationimage = '';
-        $presentation = (array)json_decode($certification->presentationjson);
-        if (!empty($presentation['image'])) {
-            $imageurl = \core\url::make_file_url(
-                "$CFG->wwwroot/pluginfile.php",
-                '/' . $context->id . '/tool_mucertify/image/' . $certification->id . '/' . $presentation['image'],
-                false
-            );
+        $imageurl = certification::get_image_url($certification, false);
+        if ($imageurl) {
             $certificationimage = '<div class="certificationimage">' . \html_writer::img($imageurl, '') . '</div>';
         }