+++ tests/phpunit/local/certification_test.php
@@ -322,6 +322,67 @@ final class certification_test extends \advanced_testcase {
         $this->assertTrue($fs->file_exists($syscontext->id, 'tool_mucertify', 'image', $certification3->id, '/', 'otherimage.jpg'));
     }
 
+    public function test_get_image_url(): void {
+        $syscontext = \context_system::instance();
+
+        $admin = get_admin();
+        $this->setUser($admin);
+        $draftid = \file_get_unused_draft_itemid();
+        $fs = get_file_storage();
+        $context = \context_user::instance($admin->id);
+        $record = [
+            'contextid' => $context->id,
+            'component' => 'user',
+            'filearea' => 'draft',
+            'itemid' => $draftid,
+            'filepath' => '/',
+            'filename' => 'image.png',
+        ];
+        $fs->create_file_from_string($record, 'content is irrelevant');
+        $data = (object)[
+            'fullname' => 'Some certification',
+            'idnumber' => 'SP1',
+            'contextid' => $syscontext->id,
+            'image' => $draftid,
+        ];
+        $certification1 = certification::create($data);
+        $this->assertSame('{"image":"image.png"}', $certification1->presentationjson);
+
+        $data = (object)[
+            'fullname' => 'Some other certification',
+            'idnumber' => 'SP2',
+            'contextid' => $syscontext->id,
+        ];
+        $certification2 = certification::create($data);
+        $this->assertSame('[]', $certification2->presentationjson);
+
+        $this->assertSame(
+            "https://www.example.com/moodle/pluginfile.php/{$syscontext->id}/tool_mucertify/image/{$certification1->id}/image.png",
+            certification::get_image_url($certification1, false)->out(false)
+        );
+        $this->assertSame(
+            "https://www.example.com/moodle/pluginfile.php/{$syscontext->id}/tool_mucertify/image/{$certification1->id}/image.png",
+            certification::get_image_url($certification1, true)->out(false)
+        );
+
+        $this->assertSame(
+            null,
+            certification::get_image_url($certification2, false)
+        );
+        $this->assertSame(
+            "https://www.example.com/moodle/pluginfile.php/{$syscontext->id}/tool_mucertify/image/{$certification2->id}/geopattern.svg",
+            certification::get_image_url($certification2, true)->out(false)
+        );
+    }
+
+    public function test_get_image_geopattern(): void {
+        $geopattern = certification::get_image_geopattern(77);
+        $this->assertStringStartsWith(
+            '<?xml version="1.0"?><svg xmlns="http://www.w3.org/2000/svg" width="340" height="340"><rect x="0" y="0" width="100%" height="100%"',
+            $geopattern->toSVG()
+        );
+    }
+
     public function test_archive(): void {
         $syscontext = \context_system::instance();